Simplify plugin discovery and loading
We experimented with a set of schemes where we discover/load/discover plugins.

This was so the app could launch a worker for `worker start` but launch content sources
for `run`.

This made the registry.py very complicated and difficult to reason about.

Now, we have simplified registry, and then only booted the application in to just getting
the `singletons` which is all that is needed to launch the consumer workers.

Thus, we now have a **much** simpler `registry.py`.